backend/ml_api/requirements.txt
# Copyright 2025, Battelle Energy Alliance, LLC, ALL RIGHTS RESERVED

annotated-types==0.7.0
anyio==4.7.0
black==24.10.0
boto3==1.37.8
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.6
filelock==3.16.1
fsspec==2024.10.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.0
openai==1.57.2
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==13.9.4
rich-toolkit==0.12.0
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==78.1.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.2
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.7.0
tqdm==4.67.1
transformers==4.52.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.32.1
watchfiles==1.0.3
websockets==14.1
